fix error

+++ src/main/java/com/example/movie/entity/Movie.java
@@ -6,7 +6,6 @@ import jakarta.persistence.Id;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.awt.*;
 import java.io.Serializable;
 import java.time.LocalDateTime;
 
@@ -17,7 +16,7 @@ import java.time.LocalDateTime;
 public class Movie implements Serializable {
 
     @Id
-    private Long id;
+    private Integer id;
     private String title;
     private String description;
     private Float rating;
